requirements.txt
flask
flask-wtf
flask-sqlalchemy
flask-bcrypt
flask-login
Pillow
flask-mail
pymysql
gunicorn
pytz
python-dotenv